backend/requirements.txt
Flask==2.1.0
Flask-RESTful==0.3.9
firebase-admin==5.3.0
requests==2.27.1
python-dotenv==0.19.2
Werkzeug==2.0.3
